==> Assay/02.ImmuneHubGenes/01.ImmuneGenes.Rmd <==
---
title: "Immune-related Differential Expression Genes"
date: "Created: 2021-08-17 Updated: `r Sys.Date()`"
output: 
  html_notebook:
    codes: hide
---

### Introdution
```{r setup, include=FALSE}
knitr::opts_chunk$set(warning = F, message = F)
library(dplyr)
library(tibble)
library(convert)
library(data.table)
library(ggplot2)
library(clusterProfiler)
library(enrichplot)
library(patchwork)

# rm(list = ls())
options(stringsAsFactors = F)
options(future.globals.maxSize = 1000 * 1024^2)

grp <- c("Normal", "Tumor")
grp.col <- c("#568875", "#73FAFC")
```


### Immune genes from ImmPort and InnateDB

Downloading the data from two websites and the data is in *../Study/ImmuneRelatedGenes*.

```bash
perl get_ImmuneRelatedGenes.pl -f1 Immport_GeneList.tsv -f2 Innatedb_curated_genes.tsv -f3 gene_mouse2human.tsv -o ImmuneRelatedGenes.tsv
```

Finally, we acquired the immune-related genes

### Importing Data 
```{r}
HNSC_ExprSet <- readRDS("../../Study/SeqData/TCGA_HNSC/Clean/TCGA-HNSC-post_mRNA_ExprSet.RDS")
HNSC_DEG <- fread("../../Result/Differential/HNSC_limma_gene.csv")
Immune_genes <- fread("../../Study/ImmuneRelatedGenes/ImmuneRelatedGenes.tsv")
```


### Immune-relted genes
```{r}
get_immnue_genes <- function(dataset=HNSC_ExprSet,
                             dataDEG=HNSC_DEG,
                             genelist=Immune_genes){

  # dataset=HNSC_ExprSet
  # dataDEG=HNSC_DEG
  # genelist=Immune_genes  
  
  diff_gene <- dataDEG %>% filter(Enrichment != "Nonsignif")
  geneid <- intersect(diff_gene$GeneID, genelist$GeneID)
  diff_gene_immune <- diff_gene %>% filter(GeneID%in%geneid)
  print(table(diff_gene_immune$Enrichment))
  
  phenotype <- pData(dataset)
  profile <- exprs(dataset) %>% data.frame() %>%
    rownames_to_column("GeneID") %>%
    filter(GeneID%in%geneid) %>%
    column_to_rownames("GeneID")

  if(!any(colnames(profile) == rownames(phenotype))){
    stop("The order of samplenames between phenotype and profile was wrong")
  }
  
  exprs <- as.matrix(profile)
  adf <-  new("AnnotatedDataFrame", data=phenotype)
  experimentData <- new("MIAME",
          name="Hua Zou", lab="gdl Lab",
          contact="zouhua1@outlook.com",
          title="Tumor Experiment",
          abstract="ExpressionSet",
          url="www.zouhua.top",
          other=list(notes="The Immune-related genes ExpressionSet"))
  expressionSet <- new("ExpressionSet", exprs=exprs,
                         phenoData=adf,
                         experimentData=experimentData)

  res <- list(immune=diff_gene_immune,
              ExprSet=expressionSet)
  
  return(res)  
  
}

if(!dir.exists("../../Result/Profile")){
  dir.create("../../Result/Profile", recursive = T)
}

HNSC_immune <- get_immnue_genes(dataset=HNSC_ExprSet,
                 dataDEG=HNSC_DEG,
                 genelist=Immune_genes)
HNSC_immune$ExprSet
saveRDS(HNSC_immune$ExprSet, "../../Result/Profile/HNSC_ExprSet_ImmuneGene.RDS", compress = TRUE)
write.csv(HNSC_immune$immune, "../../Result/Differential/HNSC_limma_ImmuneGene.csv", row.names = F)
```


### GO and KEGG pathyways enrichment analysis
```{r, fig.width=12, fig.height=7}
get_ORA <- function(dataset=HNSC_immune$immune,
                    group_name=grp,
                    type="KEGG",
                    GOType="CC"){

  # dataset=HNSC_immune$immune
  # group_name=grp
  # type="KEGG"
  # GOType="CC"

  # extract DEGs
  dat <- dataset %>%
    dplyr::rename(SYMBOL=GeneID, log2FoldChange=logFC) %>%
    filter(Enrichment%in%group_name) %>%
    distinct() %>%
    arrange(log2FoldChange)

  # transform geneid
  require(org.Hs.eg.db)
  symbol2entrz <- AnnotationDbi::select(org.Hs.eg.db,
                                      key=dat$SYMBOL,
                                      columns="ENTREZID",
                                      keytype="SYMBOL")
  symbol2entrz <- as_tibble(symbol2entrz)

  dat_cln <- inner_join(symbol2entrz, dat, by ="SYMBOL") %>%
    dplyr::select(SYMBOL, ENTREZID, log2FoldChange)
  mdat <- dat_cln %>%
    dplyr::select(ENTREZID, log2FoldChange) %>%
    na.omit() %>%
    distinct() %>%
    group_by(ENTREZID) %>%
    summarize(log2FoldChange=mean(log2FoldChange)) %>%
    ungroup() %>%
    arrange(desc(log2FoldChange))
  geneList <- tibble::deframe(mdat)

  if(type == "GO"){
    #ont MF:Molecular Function; CC: cellular Component; BP:Biological Process.
    es_res <- enrichGO(gene     = names(geneList),
                       OrgDb    = org.Hs.eg.db,
                       keyType  = "ENTREZID",
                       ont      = GOType,
                       pAdjustMethod = "BH",
                       pvalueCutoff  = 0.05,
                       qvalueCutoff  = 0.05)
  }else if(type == "KEGG"){
    es_res <- enrichKEGG(gene     = names(geneList),
                         organism = "hsa",
                         keyType  = "kegg",
                         pvalueCutoff  = 0.05,
                         pAdjustMethod = "BH",
                         qvalueCutoff  = 0.05)
  }


  res <- list(glist=geneList, es=es_res, dat=dat_cln)

  return(res)
}

plot_Escore <- function(dataset=HNSC_KEGG,
                        number=10){
  # dataset=HNSC_KEGG
  # number=10

  edox <- setReadable(dataset$es, 'org.Hs.eg.db', 'ENTREZID')
  pl <- barplot(edox, showCategory = number)+
            theme(axis.title = element_text(face = 'bold',color = 'black',size = 14),
                  axis.text = element_text(color = 'black',size = 10),
                  text = element_text(size = 8, color = "black", family="serif"),
                  legend.position = 'right',
                  legend.key.height = unit(0.6,'cm'),
                  legend.text = element_text(face = "bold", color = 'black',size = 10),
                  strip.text = element_text(face = "bold", size = 14))

  return(pl)
}

# KEGG 
HNSC_KEGG <- get_ORA(dataset=HNSC_immune$immune, type="KEGG")
table(HNSC_KEGG$es@result$p.adjust < 0.01)
HNSC_KEGG_dotplot <- plot_Escore(dataset=HNSC_KEGG, number=10)
HNSC_KEGG_dotplot

# GO terms
HNSC_GOMF <- get_ORA(dataset=HNSC_immune$immune, type="GO", GOType = "MF")
table(HNSC_GOMF$es@result$p.adjust < 0.01)
HNSC_GOMF_barplot <- plot_Escore(dataset = HNSC_GOMF, number = 10)
HNSC_GOMF_barplot

HNSC_GOCC <- get_ORA(dataset=HNSC_immune$immune, type="GO", GOType = "CC")
#table(HNSC_GOCC$es@result$p.adjust < 0.01)
HNSC_GOCC_barplot <- plot_Escore(dataset = HNSC_GOCC, number = 10)
HNSC_GOCC_barplot

HNSC_GOBP <- get_ORA(dataset=HNSC_immune$immune, type="GO", GOType = "BP")
HNSC_GOBP_barplot <- plot_Escore(dataset = HNSC_GOBP, number = 10)
HNSC_GOBP_barplot
```


### systemic information
```{r}
sessionInfo()
```
