==> Assay/02.ImmuneHubGenes/03.HubGenes_pathway.Rmd <==
---
title: "survival-associated immune-related hub genes"
date: "Created: 2021-09-14 Updated: `r Sys.Date()`"
output: 
  html_notebook:
    codes: hide
---

### Introdution
```{r setup, include=FALSE}
knitr::opts_chunk$set(warning = F, message = F)
library(dplyr)
library(tibble)
library(convert)
library(ggplot2)
library(clusterProfiler)
library(enrichplot)
library(patchwork)

# rm(list = ls())
options(stringsAsFactors = F)
options(future.globals.maxSize = 1000 * 1024^2)

grp <- c("Normal", "Tumor")
grp.col <- c("#568875", "#73FAFC")
```


### Importing data 
```{r}
Hubgene_brown <- readRDS("../../Result/WGCNA/HNSC_WGCNA_ImmuneGene_HubGene_brown.RDS")
Hubgene_yellow <- readRDS("../../Result/WGCNA/HNSC_WGCNA_ImmuneGene_HubGene_yellow.RDS")
```


### Gene Set Enrichment Analysis 
```{r}
get_GSEA <- function(dataset=Hubgene_brown,
                     type="KEGG",
                     ntop=10){

  # dataset=Hubgene_brown
  # type="KEGG"
  # ntop=10
  
   # extract DEGs
  dat <- dataset$hubGenes %>%
    dplyr::rename(SYMBOL=feature, GS=gs_value) %>%
    distinct() %>%
    arrange(GS)

  # transform geneid
  require(org.Hs.eg.db)
  symbol2entrz <- AnnotationDbi::select(org.Hs.eg.db,
                                      key=dat$SYMBOL,
                                      columns="ENTREZID",
                                      keytype="SYMBOL")
  symbol2entrz <- as_tibble(symbol2entrz)

  dat_cln <- inner_join(symbol2entrz, dat, by ="SYMBOL") %>%
    dplyr::select(SYMBOL, ENTREZID, GS)

  mdat <- dat_cln %>%
    dplyr::select(ENTREZID, GS) %>%
    na.omit() %>%
    distinct() %>%
    group_by(ENTREZID) %>%
    summarize(GS=mean(GS)) %>%
    ungroup() %>%
    arrange(desc(GS))
  geneList <- tibble::deframe(mdat)
  
  # Fold Enrichment
  get_EF <- function(ratio){
      sapply(ratio, function(x) as.numeric(strsplit(x,"/")[[1]][1])/as.numeric(strsplit(x,"/")[[1]][2]))
    }   

  if(type == "GO"){
    #ont MF:Molecular Function; CC: cellular Component; BP:Biological Process.
    es_res <- enrichGO(gene     = names(geneList),
                       OrgDb    = org.Hs.eg.db,
                       keyType  = "ENTREZID",
                       ont      = "All",
                       pAdjustMethod = "BH",
                       pvalueCutoff  = 0.05,
                       qvalueCutoff  = 0.05)
    
    df_es <- data.frame(es_res) %>%
      group_by(ONTOLOGY) %>%
      dplyr::slice(1:ntop) %>%
      ungroup() %>%
      mutate(ONTOLOGY=factor(ONTOLOGY, levels = c("BP", "CC", "MF"),
                             labels = c("Biological Process", "Cellular Componet", "Molecular Function"))) %>%
      arrange(ONTOLOGY, desc(Count))
    enrichment_fold <- as.numeric(get_EF(df_es$GeneRatio)/get_EF(df_es$BgRatio)) 
    df_es$enrichment_fold <- enrichment_fold  
    
    df_es$Description <- factor(df_es$Description, levels = as.character(df_es$Description))
    
    pl <- ggplot(data = df_es, aes(x=Description, y=Count, fill=ONTOLOGY))+
      geom_bar(stat = "identity", width = 0.8, color = "black")+
      scale_fill_manual(values = c("#E860A4", "#9748BC", "#5AADD8"))+
      labs(x="GO term", y="Num of Genes")+
      scale_y_continuous(expand = c(0, 0))+
      coord_flip()+
      theme_bw()+
      theme(axis.title = element_text(face = 'bold',color = 'black',size = 14),
            axis.text = element_text(color = 'black', size = 10),
            text = element_text(size = 8, color = "black", family="serif"),
            legend.position = c(.9, 0.2),
            legend.justification = c(0.9, 0.2),
            legend.key.height = unit(0.6,'cm'),
            legend.text = element_text(face = "bold", color = 'black',size = 10),
            strip.text = element_text(face = "bold", size = 14))
    
  }else if(type == "KEGG"){
    es_res <- enrichKEGG(gene     = names(geneList),
                         organism = "hsa",
                         keyType  = "kegg",
                         pvalueCutoff  = 0.05,
                         pAdjustMethod = "BH",
                         qvalueCutoff  = 0.05)

    df_es <- data.frame(es_res) %>%
      dplyr::slice(1:ntop)
    enrichment_fold <- as.numeric(get_EF(df_es$GeneRatio)/get_EF(df_es$BgRatio)) 
    df_es$enrichment_fold <- enrichment_fold
    
    plotdata <- df_es %>% arrange(desc(enrichment_fold), p.adjust)
    
    pl <- ggplot(data = plotdata, aes(x=enrichment_fold, y=Description))+
      geom_point(aes(color=-log10(p.adjust), size=Count))+
      scale_colour_gradient(high = 'red',low = 'green')+
      labs(x="Fold Enrichment", y="KEGG Pathway")+
      theme_bw()+
      theme(axis.title = element_text(face = 'bold',color = 'black',size = 14),
            axis.text = element_text(color = 'black', size = 10),
            text = element_text(size = 8, color = "black", family="serif"),
            legend.key.height = unit(0.6,'cm'),
            legend.text = element_text(face = "bold", color = 'black',size = 10),
            strip.text = element_text(face = "bold", size = 14))    
      
  }
  
  return(pl)
}
```


### brown modules
```{r}
get_GSEA(dataset=Hubgene_brown, type="KEGG")
get_GSEA(dataset=Hubgene_brown, type="GO")
```


### brown modules
```{r}
get_GSEA(dataset=Hubgene_yellow, type="KEGG")
get_GSEA(dataset=Hubgene_yellow, type="GO")
```


### systemic information
```{r}
sessionInfo()
```


### Reference

1. [Maximally Selected Rank and Statistics](https://www.rdocumentation.org/packages/maxstat/versions/0.7-24/topics/maxstat.test)

2. [Determine optimal cutpoints for numerical variables in survival plots](http://r-addict.com/2016/11/21/Optimal-Cutpoint-maxstat.html)
